Treat ETIMEDOUT as indicating a non-recoverable connection failure.
When the TCP_USER_TIMEOUT socket option is defined on a connection and
the timeout is reached, it will cause a read to return ETIMEDOUT. After
that the socket will be closed by the kernel. This considers ETIMEDOUT
an unrecoverable connection failure when TCP_USER_TIMEOUT is defined.

It specifically does not do make this change when WIN32 is defined. This
is to reduce the chance of introducing issues and also so that
win32_port.h does not have to be changed in an ABI incompatible way.