Patched CreateRestartPoint to reproduce invalid checkpoint at startup.
Sleep before checking ControlFile->state at CreateRestartPoint, to
increase the chance it changed to DB_IN_PRODUCTION by promotion.
If that happens, the ControlFile->checkPoint is stale and invalid.
If that was a fast promotion, the post-recovery checkpoint can be
skipped if crash happens, then the database will not recover.

Also added chaos_sigsegv to simulate a crash, and some logs to
show the timing of restartpoint, promotion and the crash.