Restructure ALTER TABLE notes to clarify table rewrites and verification scans
The current set of notes includes several examples of when rewriting table
(and its indexes) or scanning a table to validate constraints is required
and when it can be avoided. However the structure makes this hard to
follow. Here we layout out rewriting and validating as two distinct
possibilities and list the cases where they happen. That also has the
benefit of moving some of the details about rewrites (e.g., time and
disk space requirements) out of the discussion for a specific operation.

Thanks to David G. Johnston for the initial inspiration for this
structuring.