Allow building only trusted or untrusted PL/Tcl.
Presently, when the --with-tcl configuration option is used, both
trusted and untrusted PL/Tcl are built.  However, some users may
only want to build one or the other.  This change introduces an
optional argument that can be used to do so.  If
--with-tcl='trusted' is specified, only trusted PL/Tcl is built.
If --with-tcl='untrusted' is specified, only untrusted PL/Tcl is
built.  If --with-tcl is given without an argument, both trusted
and untrusted PL/Tcl are built.