Add signal-safe function to write to log
Any functions that use `malloc` directly or indirectly are not
signal-safe, and `vfprintf` is not signal-safe but used from inside
`bgworker_die`. This can cause the system to end up in a deadlock and
not respond to signals.

This commit adds a signal-safe version of write_stderr() and uses it
from signal handlers rather than write_stderr().