doc: clarify example of serialization anomaly
This example has been pointed out as confusing in multiple places, e.g.:

 * https://postgrespro.com/list/thread-id/1845118
 * https://dba.stackexchange.com/a/43951
 * https://stackoverflow.com/a/50733640 ("this paragraph is dark")

As mentioned in those links, it seems to attempt to summarize the wiki's
example at https://wiki.postgresql.org/wiki/SSI#Deposit_Report . I'm not
sure if/how to link to the wiki from the docs, so try again to summarize
it concisely and (hopefully) clearly.

Alternatively, this example could be removed in favor of a reference to
the Serializable section and its example.