WIP: Reduce memory context overheads
Whenever we palloc a chunk of memory, traditionally, we prefix the
returned pointer with a pointer to the memory context which the chunk
belongs to.  This is required so that we're able to easily determine the
owning context when performing operations such as pfree() and repalloc().
The AllocSet context additionally prefixes the pointer to the owning
context with the size of the chunk which makes the header 16 bytes in
size.

For generation contexts, the problem was worse.  In addition to the
pointer to the owning context, we also stored a pointer to the owning
block then prefixed the chunk size, which resulted in a 24-byte header.

Here we completely change the rule that pointers must be directly prefixed
by the owning memory context and instead, insist that they're directly
prefixed by a uint64 variable where the least significant 3-bits are set
to a value to indicate which type of memory context the pointer belongs
to. We're then able to pass the pointer to the function specific to that
context implementation to allow it to determine the actual memory context
to which the pointer belongs.

Here we implement a general-purpose way of encoding the remaining bits in
the prefixing uint64 value which allows encoding of the chunk size and
also the number of bytes to subtract from the pointer to get a pointer to
the block the pointer belongs to.  All memory contexts now subtract the
block offset bytes to obtain a pointer to the block and that block
contains a pointer to the owning context.  This means we only store a
pointer to the context once per block, rather than once per chunk as we
did previously.

Because the remaining space in the prefixing uint64 value after storing
the 3-bit value to indicate the memory context type is only 61 bits, we
only have 30 bits each to store the chunk size and the block offset
values.  We do allow much larger allocations than what can be stored in 30
bits, so we also have a concept of a "large" chunk and we mark if the
chunk is small or large with the 1 remaining bit in the uint64 header.

When a chunk is large, the uint64 header value is immediately preceded by
two Size values, one to indicate the chunk size and one for the block
offset.  For MEMORY_CONTEXT_CHECKING builds, each chunk header is prefixed
by the number of bytes in the memory request.  This allows us to perform
all of the memory context checking still.