Implement dual-pivot quicksort
Choose pivots by running insertion sort on five candidates and choosing
the 2nd and 4th, ("tertile of five"). If any of the five are equal, we
assume the input has many duplicates and fall back to B&M since it's
optimized for that case.

Set default insertion sort threshold to 12, since it seems better
for dual pivot.