CorrectiveQuals is as simple as a List of RestrictInfo, a). only one restrictinfo on this group should be counted for any joinrel estimation. b). at least 1 restrictinfo in this group should be executed during execution. In this commit, only rows estimation issue is addressed.
PlannerInfo.correlative_quals is added to manage all the CorrectiveQuals at
subquery level. RelOptInfo.cqual_indexes is a List * to indicate a which
CorrectiveQuals this relation related to. This is designed for easy to check if
the both sides of joinrel correlated to the same CorrectiveQuals. Why isn't the
type a Bitmapset * will be explained later.

The overall design of handing the joinrel size estimation is:
a). At the base relation level, we just count everything with the correlative
quals. b). During the any level joinrel size estimation, we just keep 1 side's
cqual (short for corrective qual) selectivity by eliminated the other one. so
the size estimation for a mergeable join selectivity becomes to:

rows = R1.rows X r2.rows X 1 / Max (ndistval_of_colA,  ndistinval_of_colB) X 1 /
Selectivity(R1's CorrectiveQual).

r1.rows X 1 / Selectivity(R1's CorrectiveQual) eliminated the impact of
CorrectiveQual on R1. After this,  the JoinRel of (R1, R2) still be impacted by
this CorrectiveQual but just one in this level.  Later if JoinRel(R1, R2) needs
to join with R3,  and R3 is impacted by this CorectiveQuals as well. This we
need to keep one and eliminating the other one as above again.

The algorithm for which Selectivity should be eliminated and which one should be
kept is:

When we join 2 inner_rel and outer_rel with a mergeable join restrictinfo, if
both sides is impacted with the same CorrectiveQual, we first choose which "side"
to eliminating based on which side of the restrictinfo has a higher distinct
value. The reason for this is more or less because we used "Max"(ndistinctValT1,
ndistinctValT2). After decide which "side" to eliminating, the real eliminating
selecitity is the side of RelOptInfo->cqual_selectivity[n]

Selectivity *RelOptInfo->cqual_selectivity:

 The number of elements in cqual_selecitity equals
 the length of cqual_indexes. The semantics is which
 selectivity in the corresponding CorectiveQuals's qual
 list is taking effect. At only time, only 1 Qual
 Selectivity is counted for any-level of joinrel.

and the other side's RelOptInfo->cqual_selectivty is used to set the upper
joinrel->cqual_selecivity.

In reality, it is possible for to have many CorrectiveQuals, but for design
discussion, the current implementation only take care of the 1 CorrectiveQuals.
this would be helpful for PoC/review/discussion.

Some flow for the key data:

1. root->corrective_quals is initialized at
generate_base_implied_equalities_no_const stage. we create a CorrectiveQual in
this list for each ec_filter and fill the RestrictInfo part for this cqual. At
the same time, we note which RelOptInfo (cqual_indexes) has related to this cqual.

2. RelOptInfo->cqual_selecitity for baserel is set at the end of set_rel_size,
at this time, the selectivity for every RestrictInfo is calcuated, we can just
fetch the cached value.  As for joinrel, it is maintained in
calc_join_cqual_selectivity, this function would return the Selectivity to
eliminate and set the above value.

Limitation in this PoC:
1. Only support 1 CorrectiveQual in root->correlative_quals
2. Only tested with INNER_JOIN.
3. Inherited table is not supported.