Fix two tabs are regarded as auto completion in test case
In this test case, when I paste the CREATE TABLE into psql,
two tabs between `name` and `text`, which are originally used for
seperation, are regarded as auto completion instruction by psql,
causing syntax error because there is no more seperation between
column name and type. Would it be better to use space here, although
the tabs are not breaking the regression?