Adjust ndistinct with nrows in the rel when estimating join selectivity.
Ndistinct is key to the accuracy of join selectivity estimation, which
impacts the performance of complex SQLs a lot. Previous code in eqjoinsel
does not take rel's restriction into account. A good math model should
use the dependency of the Vars in rel's restrictions and the join var to
estimate Ndistinct. At least there is a truth that Ndistinct should not
be greater then the number of rows of the rel.