Generalize ereport_startup_progress infrastructure
The ereport_startup_progress infrastructure added by commit
9ce346e will be super-useful for reporting progress of any
long-running server operations, not just the startup process
operations. For instance, postmaster can use it for reporting
progress of temp file and temp relation file removals,
checkpointer can use it for reporting progress of snapshot or
mapping file processing or even WAL file processing and so on.
And I'm sure there can be many places in the code where we have
while or for loops which can, at times, take a long time to finish
and having a log message there would definitely help.

This patch attempts to generalize the ereport_startup_progress
infrastructure.