pg_get_constraintdef: Schema qualify foreign tables by default
This matches pg_get_constraintdef to behave the same way as
pg_get_indexdef, which is to schema qualify referenced objects in the
definition, unless pretty printing is explicitly requested.

For pretty printing the previous behaviour is retained, which is to only
include the schema information if the referenced object is not in the
current search path.