initdb: do not exit when PGDATA, PGDATA/pg_wal is a mountpoint
With the rise of containerized environments postgresql is running at,
more people are using to mount the persistent storage devices over the PGDATA
path (e.g. `/var/lib/postgresql/data`).

With that, the `lost+found` directory gets created in that path.

According to already existing warn_on_mount_point() for errno 2 & 3 - it is
expected only to warn the user when it assumes the directory being is
a mount point.

This patch eliminates initdb to exit when it finds the PGDATA is a mountpoint,
which is now often the case.