Avoid distributing new catalogsnapshot for the transaction being decoded
Currently, when decoding the COMMIT of a transaction containing catalog
changes, we add a new snapshot to all transactions including the one being
decoded. But since we've already built a new snapshot for the current
transaction, there's no need to add a new one. So skip adding a snapshot of the
current transaction being decoded to improve this.