Move postmaster's state out of global variables.
Historically, the postmaster's state for tracking its children had to
live in global variables, because it had to be accessible from signal
handlers.

After commit 7389aad6, we can shift most of it into a struct that is
allocated on the heap, and freed in child processes.

This is a mechanical change: variable names remain the same, they're
just moved into a Postmaster struct and accessed as pm->X rather than X,
with pm passed down to relevant subroutines.

XXX This is not complete.  There is more postmaster state in bgworker.c.
The problem is that BackgroundWorkerList probably belongs in Postmaster
struct, as far as tracking running bgworkers goes, but it's also
accessed by RegisterBackgroundWorker(), called by extensions, to
register bgworkers that should be running.