Ignore dropped columns when checking the column list in logical replication
There's a restriction that different column lists for same table can't be used
in the publications of single subscription. But in the check of walsender and
table synchronization, the dropped columns are included when getting the number
of columns in a table. This can lead to unexpected errors in some cases. Fix it
in this patch.