Aggregate push-down - basic functionality.
With this patch, partial aggregation can be applied to a base relation or to a
join, and the resulting "grouped" relations can be joined to other "plain"
relations. Once all tables are joined, the aggregation is finalized. See
README for more information.

The next patches will enable the aggregate push-down feature for parallel
query processing, for partitioned tables and for foreign tables.