btree specialization: Prepare code for specialization.
nbtree index keys are not all made the same, and a significant amount of
time is spent in code that exists only to deal with various key shapes. By
allowing the specialization of functions based on the key shape, we can
remove or reduce these causes of overhead.

This commit moves the various specializable functions into "_spec"
sourcefiles. Later commits will further modify these sources, mostly
mechanically, in a way that produces several optimized variants of the
functions for several distinct index key shapes.