Rework LogicalOutputPluginWriterUpdateProgressAndKeepalive
Currently whenever we need to add a callback to the output plugin, the
output plugin needs to confirm whether it needs to call
WalSndUpdateProgressAndKeepalive or not. As the number of callbacks
increases, this framework becomes less maintainable. So we decided to
improve this by calling WalSndUpdateProgressAndKeepalive from the
common code, rather than in the output plugins.

We introduced LogicalDecodingContext->did_write to confirm whether the
output plugins actually sends data or not. We use this flag to check
if we should send a keepalive message to the subscriber prevent delays,
when an empty transaction is skipped in synchronous logical replication.

Since we can know whether the change is an end of transaction change in
the common code, we removed the LogicalDecodingContext->end_xact
introduced in commit f95d53e.

In commit 8c58624, in order to fix logical replication timeout during
large DDLs, we tried to send the keepalive message to the subscriber for
all change types in the function ReorderBufferProcessTXN. Now, we only
handle the timeout-related issues in the common change type (see function
is_keepalive_threshold_exceeded).