Don't crash on out of memory when starting backend
There are no checks for the return values of several allocations using
malloc(), calloc(), and strdup() in backend startup which means that
it is possible to get a segmentation fault when starting a backend if
the system is low on memory. It makes more sense to abort the spawning
of a backend if we're low on memory rather than tearing down the
entire server.

This commit fixes this by using memory context allocations rather than
malloc and friends.

Since the postmaster should not exit if a memory allocation failure
occurs when trying to start a backend, it will use the no-error version
of the memory context allocation functions and add checks if we're out
of memory, log an error, and just continue running.

Memory allocation errors when initializing the backend process will
generate errors instead and as a result send the error message to the
connecting client indicating that the server is out of memory and the
backend will shut down after reporting the error.