Implement the USAGE privilege on PUBLICATION.
Publication row filters and column lists can be used to prevent subsets of
data from being replicated via the logical replication system. These features
can address performance issues, but currently should not be used for security
purposes, such as hiding sensitive data from the subscribers. The problem is
that any subscriber can get the data from any publication, so if the sensitive
data is deliberately not published via one publication, it can still be
available via another one (supposedly created for another subscriber).

This patch adds an ACL column to the pg_publication catalog, implements the
corresponding checks and enhances the GRANT and REVOKE commands to grant and
revoke the USAGE privilege on publication to / from roles. The USAGE privilege
is initially granted to the PUBLIC group (so that existing configurations
don't get broken) but the user can revoke it and grant it only to individual
subscription users (i.e. users mentioned in the subscription connection
configuration). Thus the publisher instance can reject to send data that given
subscriber is not supposed to receive.

Obviously, the publication privileges are checked on the publisher side,
otherwise the implementation wouldn't be secure. The output plugin
(pgoutput.c) is the easy part because it already does receive the list of
publications whose data it should send to the subscriber. The initial table
synchronization is a little bit tricky because so far the "tablesync worker"
(running on the subscriber side) was responsible for constructing the SQL
query for the COPY TO command, which is executed on the publisher side.

This patch adds a new option PUBLICATION_NAMES to the COPY TO command. The
subscriber uses it to pass a list of publications to the publisher. The
publisher checks if the subscription user has the USAGE privilege on each
publication, retrieves the corresponding data (i.e. rows matching the row
filters of the publications) and sends it to the subscriber.

Since the publisher and subscriber instances can be on different major
versions of postgres, and since old subscribers cannot send the publication
names during the initial table synchronization, a new configuration variable
"publication_security" was added. The default value is "off", meaning that the
publisher does not require the COPY TO command to contain the
PUBLICATION_NAMES option. If the option is passed yet, the publisher does not
check the privileges on the listed publications, but it does perform row
filtering according to the publication filters. Thus upgrade of the publisher
instance does not break anything.

Once all the subscribers have migrated to the postgres version that supports
this feature, this variable should be set to "on". At that moment the
publisher starts to require the presence of the PUBLICATION_NAMES option in
the COPY TO command, as long as the COPY TO is executed by a role which has
the REPLICATION privilege. (Role w/o the REPLICATION privilege aren't
currently allowed to use the PUBLICATION_NAMES option.)