Make autovacuum docs into a sect1 of its own.
This doesn't change any of the content itself.  Though it does move it
from the end of "Routine Vacuuming" (which is itself a sect1) to a whole
new sect1 that appears _before_ "Routine Vacuuming".

This commit is as mechanical as possible, in the sense that it is
structured in such a way as to make git diff's "dimmed-zebra" mode show
as few true changes as possible (almost everything is strictly movement
of existing content).  Note in particular that this commit does not fix
any indentation (we keep sect2 indentation for our new sect1).  That
will happen in a later commit, after we're done splitting up content
from "The Autovacuum Daemon" into more subsections to improve its
readability.

XXX Open question: does it make more sense to move the sect1 to before
"Routine Vacuuming", or should it go after instead?  There are arguments
for both.

Arguments for "before" (which is how it's done by this commit right
now):

"Before" gives greater prominence to the autovacuum scheduling tunables,
such as autovacuum_vacuum_scale_factor.  These are the most important
individual tunables, which argues for putting them earlier than "Routine
Vacuuming".

Arguments for "after":

Although the discussion in "Routine Vacuuming" is rather involved, it is
arguably still introductory material that informs how the user will tune
autovacuum_vacuum_scale_factor.  (Assuming that the user doesn't just go
by trial and error, which seems more likely in practice but not
necessarily the most useful working assumption for our purposes.)